requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic-settings>=2.8,<3.0
python-dotenv>=1.0,<2.0
pipecat-ai[anthropic,cartesia,deepgram,silero,websocket]>=0.0.105,<1.0
